+++ MOUP/MOUP/Domain/Entities/TodayRoutine.swift
@@ -1,0 +1,13 @@
+//
+//  TodayRoutine.swift
+//  MOUP
+//
+//  Created by 송규섭 on 10/5/25.
+//
+
+import Foundation
+
+struct TodayRoutine {
+    let workplaceName: String
+    let routines: [Routine]
+}
