+++ MOUP/MOUP/Domain/Entities/TodoItem.swift
@@ -1,0 +1,21 @@
+//
+//  TodoItem.swift
+//  MOUP
+//
+//  Created by 송규섭 on 10/6/25.
+//
+
+import Foundation
+
+struct TodoItem: Hashable {
+    let id: UUID = UUID()
+    var text: String
+
+    func hash(into hasher: inout Hasher) {
+        hasher.combine(id)
+    }
+
+    static func ==(lhs: TodoItem, rhs: TodoItem) -> Bool {
+        return lhs.id == rhs.id
+    }
+}
