+++ MOUP/MOUP/Presentation/Routine/ViewModel/AddRoutineViewModel.swift
@@ -15,19 +15,6 @@ enum ValidationFocusTarget {
     case firstTodoItem
 }
 
-struct TodoItem: Hashable {
-    let id: UUID = UUID()
-    var text: String
-
-    func hash(into hasher: inout Hasher) {
-        hasher.combine(id)
-    }
-
-    static func ==(lhs: TodoItem, rhs: TodoItem) -> Bool {
-        return lhs.id == rhs.id
-    }
-}
-
 final class AddRoutineViewModel {
     
     // MARK: - Input
